stub out webinar content for easier editing for our devrel folks.
webinars current take multiple places to edit so instead centrialize it
in a collection to be used throughout the project. tbh this isn't really
'webinar' specific but probs should be something along the lines of
'banners.json' or another notification wide system we want in the docs
i.e. confs, new releases, webinars, etc. and leverage the `tags` or
`type` field to make selections as needed for different UI elements.