fix(release): select git tag with priority order of preid, stable and rest
Currently, when release is run and there is a pre-release version, it is used as base for the new version, leading to no changes detected or version skips. This commit changes the logic of source tag selection, taking into account the preid used for the release when one is provided, and prioritising stable release tags when no preid is provided.

closed #31740