feat(docker): add docker plugin (#31634)
## Current Behavior
We do not currently have a plugin for docker to infer targets such as
- build
- run

## Expected Behavior
The Docker Plugin should remain agnostic between tech stacks, but should
allow for the containerization of projects.

It should support initially the ability to build and run images by
inferring targets on projects that contain a Dockerfile.

`nx release` should be expanded to support releasing Docker images to
registries