fix(docker): generate dockerfile compliant with package manager (#32191)
## Current Behavior
We currently generate a Dockerfile for node applications that only takes
into account npm.
This means that if the workspace is using yarn or pnpm then when we
prune the lockfile, we have an incorrect lockfile in the output folder
for the docker build.

## Expected Behavior
Ensure the generated Dockerfile is compliant for the package manager
used in the workspace