fix(core): provide hint that vite will be used when selecting React Router (#32238)
## Current Behavior
React Router requires using a Vite plugin. This means that during the
CNW flow, if a user selects `Yes` to the prompt for `React Router for
SSR`, we do not prompt them for `bundler`.
This can be confusing for users that expect to be able to select
bundler.

## Expected Behavior
Clarify during the prompt for React Router for SSR that Vite will be
selected as the bundler.