cleanup(angular): remove CommonModule from component generators (#32215)
## Current Behavior
When generating a new Angular component it imports `CommonModule` by
default.

## Expected Behavior
The `CommonModule` is not needed as often in modern angular because of
the new control flow syntax and the move away from `[ngClass]` and
`[ngStyle]` in favour of `[class]` and `[style]`. It is also not
included by the official angular component generator.