fix(vite): set name property for test to project name #32163 (#32308)
## Current Behavior
The `name` property is not set for vitest configs. Vitest will therefore
infer the name for the test suite based off the directory name.
In a monorepo this can cause issues as the directory name may be
repeated but for different contexts.

## Expected Behavior
Set the project name as the name in the test config

## Related Issue(s)

Fixes #32163