fix(module-federation): skip splitChunk options for prod build (#32330)
## Current Behavior
Rspack + React + Module Federation applications are not able to work
properly due to the presence of splitChunks optimization on default web
configs.

## Expected Behavior
Optimizations like this should be overridden by the Module Federation
plugin in order for them to work.

## Related Issue(s)
https://github.com/nrwl/nx/issues/32329

Fixes #
Added override of splitChunks configs for Nx's Module Federation
plugins.

---------

Co-authored-by: Colum Ferry <cferry09@gmail.com>