chore(misc): remove circular dep from @nx/cypress to itself and clarify why other remaining packages are ignored (#32344)
This removes the need to ignore `@nx/cypress` in
`packages/cypress/.eslintrc.json`. This was never really a ciruclar dep,
as
`packages/cypress/src/generators/configuration/files/v9/__directory__/plugins/index.js`
is a template for `generateFiles`, thus appending `__tmpl__` fixes it.

Other packages have valid reasons so a comment is added to clarify for
the future.