fix(release): handle git_protocol https when getting auth token #31828 (#32353)
## Current Behavior
When resolving token data for creating the GitHubReleaseClient, The GH
CLI supports `https` option for `git_protocol` and it is not being
handled.
This results in no token data being resolved.

Further API requests that require a `Bearer` token in the headers fail.

## Expected Behavior
Ensure `https` protocol is also handled.
`gh auth token` still operates and produces a token that can be used.

## Related Issue(s)

Fixes #31828