Revert "feat(webpack): migrate from deprecated url.parse() to WHATWG … (#32397)
…URL API"

This reverts commit f768886ae75a3dd2ed04abc91b5ff2d9f97c3e67.

<!-- Please make sure you have read the submission guidelines before
posting an PR -->
<!--
https://github.com/nrwl/nx/blob/master/CONTRIBUTING.md#-submitting-a-pr
-->

<!-- Please make sure that your commit message follows our format -->
<!-- Example: `fix(nx): must begin with lowercase` -->

<!-- If this is a particularly complex change or feature addition, you
can request a dedicated Nx release for this pull request branch. Mention
someone from the Nx team or the `@nrwl/nx-pipelines-reviewers` and they
will confirm if the PR warrants its own release for testing purposes,
and generate it for you if appropriate. -->

## Current Behavior
<!-- This is the behavior we have today -->

Accidentally pushed this without review

## Expected Behavior
<!-- This is the behavior we should expect with the changes in this PR
-->

Reverting it so it can be reviewed

## Related Issue(s)
<!-- Please link the issue being fixed so it gets closed when this is
merged. -->

Fixes #